api/requirements.txt
Flask
Flask-Migrate
Flask-Moment
Flask-OIDC
Flask-RESTplus
Flask-Script
Flask-SQLAlchemy
cockroachdb
eventlet
flask-cors
flask-login
flask-socketio
gunicorn
kombu
psycopg2-binary
python-jose-cryptodome
